# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1"
    },
    "paths": {
...
        "/api/user-topics": {
            "post": {
                "tags": [
                    "Topics"
                ],
                "summary": "Attach topics to current user.",
                "description": "Attach topics to current user.",
                "operationId": "App\\Http\\Controllers\\TopicController::attachToUser",
                "requestBody": {
                    "description": "U can provide array of id's what shoud attached to user",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "topics": {
                                        "type": "array",
                                        "items": {},
                                        "example": [
                                            1,
                                            2,
                                            3,
                                            4
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Family"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Family description"
                                                    },
                                                    "photo": {
                                                        "type": "string",
                                                        "example": "family"
                                                    },
                                                    "created_at": {
                                                        "type": "timestamp",
                                                        "example": "2021-07-07T01:30:48.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "timestamp",
                                                        "example": "2021-07-07T01:30:48.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "title": "Family",
                                                    "description": "Family description",
                                                    "photo": "https://via.placeholder.com/640x480.png/000011?text=fuga",
                                                    "created_at": "2021-07-07T01:30:48.000000Z",
                                                    "updated_at": "2021-07-07T01:30:48.000000Z"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}